+++ CHANGES.md
@@ -3,13 +3,22 @@ To see what changed each version, look here.
 
 # Menu
 * [V1.0](#v1.0)
+  * [V1.0.4](#V1.0.4)
   * [V1.0.3](#V1.0.3)
   * [V1.0.2](#V1.0.2)
   * [V1.0.1](#v1.0.1)
   * [V1.0.0](#v1.0.0)
 
 # V1.0
-## v1.0.3
+## V1.0.4
+* Website-render won't throw errors anymore
+* Website-render only loads scripts once for each page, so you won't get errors from that anymore!
+* Website-render will now wait when the button the button is clicked and not all pages are loaded yet
+* You no longer need to put the title in the body, if it's in the head, website-render will automaticly detect it and send it to the client.
+* Added a way to show a loading icon until website-render loaded the hole page.
+* Added comments in [__frontEnd.js](src/__frontEnd.js) (Not the javascript that will be sent to the user), so everyone knows what's going on.
+* Some minor changes for the documentation.
+## V1.0.3
 * Fixed a bug with loading express routers in other files
 ## V1.0.2
 * A bug fix with the versions